Adding metrics-aws-lambda
