+++ cla-backend-go/Gopkg.lock
@@ -168,14 +168,6 @@
   version = "v0.4.0"
 
 [[projects]]
-  digest = "1:41dd6c66548c6a8e3a9bbb38132002f1907547f99246070aff172afaae51d657"
-  name = "github.com/fnproject/fdk-go"
-  packages = ["."]
-  pruneopts = "UT"
-  revision = "2ad3d2d98fa8bec0b8b1cb8f8d8c2c3a3e71bac9"
-  version = "v0.0.2"
-
-[[projects]]
   digest = "1:abeb38ade3f32a92943e5be54f55ed6d6e3b6602761d74b4aab4c9dd45c18abd"
   name = "github.com/fsnotify/fsnotify"
   packages = ["."]
@@ -571,18 +563,6 @@
   version = "v1.2.0"
 
 [[projects]]
-  branch = "master"
-  digest = "1:d414058838cbe9c12bf62ce696624ed7ef36a9b82e388d6d5ed4a20326219a78"
-  name = "github.com/tencentyun/scf-go-lib"
-  packages = [
-    "cloudfunction",
-    "cloudfunction/messages",
-    "functioncontext",
-  ]
-  pruneopts = "UT"
-  revision = "9a6ea1bf75b89a0cd917a1e846bacf40fc06ca56"
-
-[[projects]]
   digest = "1:c468422f334a6b46a19448ad59aaffdfc0a36b08fdcc1c749a0b29b6453d7e59"
   name = "github.com/valyala/bytebufferpool"
   packages = ["."]
@@ -744,7 +724,6 @@
     "github.com/awslabs/aws-lambda-go-api-proxy/httpadapter",
     "github.com/aymerick/raymond",
     "github.com/dgrijalva/jwt-go",
-    "github.com/fnproject/fdk-go",
     "github.com/go-openapi/errors",
     "github.com/go-openapi/loads",
     "github.com/go-openapi/runtime",
@@ -768,7 +747,6 @@
     "github.com/spf13/cobra",
     "github.com/spf13/viper",
     "github.com/stretchr/testify/assert",
-    "github.com/tencentyun/scf-go-lib/cloudfunction",
     "golang.org/x/net/netutil",
     "golang.org/x/oauth2",
     "golang.org/x/oauth2/github",
