+++ cla-backend-go/cmd/metrics_lambda/main.go
@@ -1,0 +1,62 @@
+package main
+
+import (
+	"context"
+	"os"
+
+	"github.com/aws/aws-lambda-go/lambda"
+
+	"github.com/aws/aws-lambda-go/events"
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/aws/session"
+	log "github.com/communitybridge/easycla/cla-backend-go/logging"
+	"github.com/communitybridge/easycla/cla-backend-go/metrics"
+)
+
+var (
+	// version the application version
+	version string
+
+	// build/Commit the application build number
+	commit string
+
+	// branch the build branch
+	branch string
+
+	// build date
+	buildDate string
+)
+
+var awsSession = session.Must(session.NewSession(&aws.Config{}))
+var metricsRepo metrics.Repository
+var stage string
+
+func init() {
+	stage = os.Getenv("STAGE")
+	if stage == "" {
+		log.Fatal("stage not set")
+	}
+	log.Infof("STAGE set to %s\n", stage)
+	metricsRepo = metrics.NewRepository(awsSession, stage)
+}
+
+func handler(ctx context.Context, event events.CloudWatchEvent) {
+	err := metricsRepo.CalculateAndSaveMetrics()
+	if err != nil {
+		log.Fatal("Unable to save metrics in dynamodb")
+	}
+}
+
+func printBuildInfo() {
+	log.Infof("Version                 : %s", version)
+	log.Infof("Git commit hash         : %s", commit)
+	log.Infof("Branch                  : %s", branch)
+	log.Infof("Build date              : %s", buildDate)
+}
+
+func main() {
+	log.Info("Lambda server starting...")
+	printBuildInfo()
+	lambda.Start(handler)
+	log.Infof("Lambda shutting down...")
+}
